refactor(problem): sync changes with backend (#78)
