+++ .changes/refactor-problem.md
@@ -1,0 +1,9 @@
+---
+"algohub": patch:refactor
+---
+
+Fully refactored the problem api to sync changes with the `algohub-server`, also:
+
+- Fixed icon sizes in markdown editor.
+- Extract supported languages from ts enum.
+- Refactored problem editor.
