+++ src/scripts/types.ts
@@ -45,8 +45,39 @@ export interface UserContent {
   id: string;
 }
 
-export interface ProblemDetail {
-  id: RecordId;
+export enum ProblemVisibility {
+  ContestOnly = "contest_only",
+  Public = "public",
+  Private = "private",
+  Internal = "internal",
+}
+
+export interface TestCase {
+  input: string;
+  output: string;
+}
+
+export interface CreateProblem {
+  id: string;
+  token: string;
+
+  title: string;
+  description: string;
+  input?: string;
+  output?: string;
+  samples: Sample[];
+  hint?: string;
+  owner: RecordId;
+  time_limit: number;
+  memory_limit: number;
+  test_cases: { input: string; output: string }[];
+  categories: string[];
+  tags: string[];
+  visibility: ProblemVisibility;
+}
+
+export interface UserProblem {
+  id: string;
   title: string;
   description: string;
   input?: string;
@@ -55,20 +86,22 @@ export interface ProblemDetail {
   hint?: string;
   time_limit: number;
   memory_limit: number;
-  test_cases: Sample[];
-  creator: RecordId;
+  test_cases: TestCase[];
+  creator: string;
   owner: RecordId;
   categories: string[];
   tags: string[];
-  mode: Mode;
-  private: boolean;
-  created_at: Date;
-  updated_at: Date;
+  visibility: ProblemVisibility;
+  created_at: string;
+  updated_at: string;
 }
 
 export enum Language {
-  Rust = "Rust",
-  Python = "Python",
-  C = "C",
-  Cpp = "Cpp",
+  Rust = "rust",
+  Python = "python",
+  C = "c",
+  Cpp = "cpp",
+  Golang = "golang",
+  Nodejs = "nodejs",
+  Java = "java",
 }
